requirements.txt
# Core
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.2.0
langchain-ollama>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Document loaders
pypdf>=4.0.0
python-docx>=1.0.0
unstructured>=0.15.0

# UI
streamlit>=1.38.0

# API
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# Fine-tuning (optional — install when needed)
# transformers>=4.44.0
# datasets>=3.0.0
# peft>=0.13.0
# trl>=0.11.0
# bitsandbytes>=0.44.0
# accelerate>=0.34.0
# torch>=2.4.0
